Feat: Modificacion de clase CItaView, integracion de menu y metodos de agregar
